requirements.txt
tensorflow>=2.10.0
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0
Pillow>=9.0.0
numpy>=1.24.0
PyQt5>=5.15.0
requests>=2.28.0
sqlite3